Fix respond_to? usage for Ruby 2.0
The path helpers are protected methods, and Ruby 2.0 changed the
behaviour of respond_to? so that it returns false for protected methods.

See http://tenderlovemaking.com/2012/09/07/protected-methods-and-ruby-2-0.html

Adding true as the second argument will check both protected and private
methods which is what we want here, given we are calling respond_to? on
self.